change result of Count to be integer
git-svn-id: http://datacenter.gradsoft.ua/repos/Jungle/jungle.public@15008 e6263924-d73c-0410-abf9-b8ad5fbdd136

+++ facilities/configuration/src/api/ua/gradsoft/jungle/configuration/ConfigurationFacade.java
@@ -13,7 +13,7 @@ public interface ConfigurationFacade
 
   List<ConfigItem>  getConfigItems(ConfigItemSelector itemSelector);
  
-  Long              getConfigItemsCount(ConfigItemSelector itemSelector);
+  Integer           getConfigItemsCount(ConfigItemSelector itemSelector);
 
   BigDecimal registerConfigItem(ConfigItem item);
